Move scripting stuff to docs/auxil
