+++ .pre-commit-config.yaml
@@ -68,7 +68,7 @@ repos:
     rev: v3.3.1
     hooks:
     -   id: pyupgrade
-        files: ^(telegram|examples|tests)/.*\.py$
+        files: ^(telegram|examples|tests|docs)/.*\.py$
         args:
           - --py37-plus
 -   repo: https://github.com/pycqa/isort
